+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/node/ProcessBesuNodeRunner.java
@@ -164,6 +164,12 @@ public class ProcessBesuNodeRunner implements BesuNodeRunner {
       params.add(Integer.toString(node.getSynchronizerConfiguration().getSyncMinimumPeerCount()));
       params.add("--Xsynchronizer-pivot-distance");
       params.add(Integer.toString(node.getSynchronizerConfiguration().getSyncPivotDistance()));
+      params.add("--Xsnapsync-synchronizer-pivot-block-window-validity");
+      params.add(
+          Integer.toString(
+              node.getSynchronizerConfiguration()
+                  .getSnapSyncConfiguration()
+                  .getPivotBlockWindowValidity()));
     } else {
       params.add("--sync-mode");
       params.add("FULL");