+++ app/src/test/java/org/hyperledger/besu/cli/options/SynchronizerOptionsTest.java
@@ -70,8 +70,6 @@ public class SynchronizerOptionsTest
             ImmutableSnapSyncConfiguration.builder()
                 .pivotBlockWindowValidity(
                     SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY + 2)
-                .pivotBlockDistanceBeforeCaching(
-                    SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_DISTANCE_BEFORE_CACHING - 2)
                 .trienodeCountPerRequest(
                     SnapSyncConfiguration.DEFAULT_TRIENODE_COUNT_PER_REQUEST + 2)
                 .storageCountPerRequest(SnapSyncConfiguration.DEFAULT_STORAGE_COUNT_PER_REQUEST + 2)