+++ ethereum/eth/src/test/java/org/hyperledger/besu/ethereum/eth/sync/common/PivotSyncActionsTest.java
@@ -44,6 +44,7 @@ import org.hyperledger.besu.ethereum.eth.peervalidation.PeerValidator;
 import org.hyperledger.besu.ethereum.eth.sync.PivotBlockSelector;
 import org.hyperledger.besu.ethereum.eth.sync.SyncMode;
 import org.hyperledger.besu.ethereum.eth.sync.SynchronizerConfiguration;
+import org.hyperledger.besu.ethereum.eth.sync.snapsync.SnapSyncConfiguration;
 import org.hyperledger.besu.ethereum.eth.sync.snapsync.SnapSyncProcessState;
 import org.hyperledger.besu.ethereum.eth.sync.state.SyncState;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
@@ -53,6 +54,7 @@ import org.hyperledger.besu.plugin.services.storage.DataStorageFormat;
 import org.hyperledger.besu.testutil.DeterministicEthScheduler;
 
 import java.io.IOException;
+import java.time.Clock;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -117,7 +119,12 @@ public class PivotSyncActionsTest {
     syncState = new SyncState(blockchain, ethPeers);
     pivotSyncActions =
         createPivotSyncActions(
-            syncConfig, new PivotSelectorFromPeers(ethContext, syncConfig, syncState));
+            syncConfig,
+            new PivotSelectorFromPeers(
+                ethContext,
+                syncConfig,
+                syncState,
+                SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY));
   }
 
   @ParameterizedTest
@@ -168,7 +175,12 @@ public class PivotSyncActionsTest {
 
     pivotSyncActions =
         createPivotSyncActions(
-            syncConfig, new PivotSelectorFromPeers(ethContext, syncConfig, syncState));
+            syncConfig,
+            new PivotSelectorFromPeers(
+                ethContext,
+                syncConfig,
+                syncState,
+                SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY));
 
     EthProtocolManagerTestUtil.createPeer(ethProtocolManager, 5000);
 
@@ -185,7 +197,12 @@ public class PivotSyncActionsTest {
     setUp(storageFormat, Optional.of(1));
     pivotSyncActions =
         createPivotSyncActions(
-            syncConfig, new PivotSelectorFromPeers(ethContext, syncConfig, syncState));
+            syncConfig,
+            new PivotSelectorFromPeers(
+                ethContext,
+                syncConfig,
+                syncState,
+                SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY));
 
     EthProtocolManagerTestUtil.createPeer(ethProtocolManager, Difficulty.of(1000), 5500);
     EthProtocolManagerTestUtil.createPeer(ethProtocolManager, Difficulty.of(2000), 4000);
@@ -203,7 +220,12 @@ public class PivotSyncActionsTest {
     setUp(storageFormat, Optional.of(2));
     pivotSyncActions =
         createPivotSyncActions(
-            syncConfig, new PivotSelectorFromPeers(ethContext, syncConfig, syncState));
+            syncConfig,
+            new PivotSelectorFromPeers(
+                ethContext,
+                syncConfig,
+                syncState,
+                SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY));
 
     EthProtocolManagerTestUtil.disableEthSchedulerAutoRun(ethProtocolManager);
 
@@ -277,7 +299,12 @@ public class PivotSyncActionsTest {
     final int peerCount = 4;
     pivotSyncActions =
         createPivotSyncActions(
-            syncConfig, new PivotSelectorFromPeers(ethContext, syncConfig, syncState));
+            syncConfig,
+            new PivotSelectorFromPeers(
+                ethContext,
+                syncConfig,
+                syncState,
+                SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY));
     final long minPivotHeight = syncConfig.getSyncPivotDistance() + 1L;
     EthProtocolManagerTestUtil.disableEthSchedulerAutoRun(ethProtocolManager);
 
@@ -326,7 +353,12 @@ public class PivotSyncActionsTest {
     setUp(storageFormat, Optional.of(1));
     pivotSyncActions =
         createPivotSyncActions(
-            syncConfig, new PivotSelectorFromPeers(ethContext, syncConfig, syncState));
+            syncConfig,
+            new PivotSelectorFromPeers(
+                ethContext,
+                syncConfig,
+                syncState,
+                SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY));
     final long pivotDistance = syncConfig.getSyncPivotDistance();
 
     EthProtocolManagerTestUtil.disableEthSchedulerAutoRun(ethProtocolManager);
@@ -387,7 +419,12 @@ public class PivotSyncActionsTest {
     setUp(storageFormat, Optional.of(1));
     pivotSyncActions =
         createPivotSyncActions(
-            syncConfig, new PivotSelectorFromPeers(ethContext, syncConfig, syncState));
+            syncConfig,
+            new PivotSelectorFromPeers(
+                ethContext,
+                syncConfig,
+                syncState,
+                SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY));
 
     final BlockHeader expectedHeader = blockchain.getBlockHeader(1).get();
     final PeerTaskExecutor peerTaskExecutor = ethContext.getPeerTaskExecutor();
@@ -428,12 +465,12 @@ public class PivotSyncActionsTest {
             syncConfig,
             new PivotSelectorFromSafeBlock(
                 blockchainSetupUtil.getProtocolContext(),
-                blockchainSetupUtil.getProtocolSchedule(),
-                ethContext,
                 genesisConfig,
-                () -> finalizedEvent,
-                () -> {},
-                headerDownloader));
+                headerDownloader,
+                blockchainSetupUtil.getProtocolSchedule(),
+                Clock.systemUTC(),
+                SnapSyncConfiguration.DEFAULT_PIVOT_BLOCK_WINDOW_VALIDITY,
+                () -> {}));
 
     final BlockHeader expectedHeader = blockchain.getBlockHeader(3).get();
     final PeerTaskExecutor peerTaskExecutor = ethContext.getPeerTaskExecutor();